requirements.txt
docker==6.1.3
Flask==3.0.0
jedi==0.19.1
keras==2.14.0
Markdown==3.5
matplotlib==3.8.0
Pillow==10.1.0
PyMuPDF==1.23.5
requests==2.31.0
scikit-learn==1.3.1
seqeval==1.2.2
simpletransformers==0.64.3
streamlit==1.27.2
tensorflow-cpu==2.14.0
transformers==4.34.1
uvicorn==0.23.2
